pyproject.toml
[tool.poetry]
name = "weather-forecast"
version = "0.0.7"
description = "Project for fastapi backend with react frontend."
authors = ["zewa-bamuga <tikhonov.igor2028@yandex.ru>"]
readme = "README.md"
packages = [{ include = "src/app" }]


[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.1"
uvicorn = "^0.30.1"
sqlalchemy = "^2.0.31"
alembic = "^1.13.2"
dependency-injector = "^4.41.0"
pydantic = "^2.8.2"
pydantic-settings = "^2.3.4"
gunicorn = "^22.0.0"
typer = "^0.12.3"
asyncpg = "^0.29.0"
psycopg2-binary = "^2.9.9"
passlib = { version = "1.7.4", extras = ["bcrypt"] }
a8t-tools = { extras = ["all"], version = "~0.1.1", source = "gitlab" }
orjson = "^3.10.6"
bcrypt = "4.0.1"
httpx = "^0.27.0"
sr = "^0.1.0"
nest-asyncio = "^1.6.0"
freezegun = "^1.5.1"
async-asgi-testclient = "^1.4.11"
factory-boy = "^3.3.0"
python-multipart = "^0.0.7"
pyjwt = {version = "2.8.0", extras = ["crypto"]}


[tool.poetry.group.dev.dependencies]
pre-commit = "3.0.4"
pytest = "^7.4.3"
pytest-asyncio = "^0.23.2"
pytest-cov = "^4.1.0"
factory-boy = "^3.3.0"
async-asgi-testclient = "^1.4.11"
pytest-timeout = "^2.2.0"
nest-asyncio = "^1.5.8"
pyfakefs = "^5.3.2"
freezegun = "^1.4.0"
flake8 = "^6.0.0"
isort = "^5.12.0"
black = "^23.1.0"
pyupgrade = "^3.3.1"
pyright = "^1.1.300"
mypy = "^1.8.0"
types-passlib = "^1.7.7.10"


[[tool.poetry.source]]
name = "gitlab"
url = "https://git.aldera-soft.ru/api/v4/projects/273/packages/pypi/simple"
priority = "primary"


[tool.pyright]
exclude = ["**/alembic", ".venv", "venv", "*/admin", "**/tests"]
pythonVersion = "3.11"
pythonPlatform = "Linux"

[tool.pytest.ini_options]
asyncio_mode = "auto"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
